setups/sprint-setup/src/requirements.txt
psycopg2-binary
structlog
httpx